+++ packages/esnext/overrides/typescript.js
@@ -27,10 +27,11 @@ module.exports = {
 
         // - Interdit l'utilisation des directives `// @ts-nocheck` et `// @ts-check` qui n'ont pas
         //   de raison d'être utilisée car seuls les fichiers `.tsx?` doivent être typés (et non les fichiers `.js`).
+        //   Pour ce qui est de `// @ts-expect-error` et `// @ts-ignore`, un commentaire est exigé.
         // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/ban-ts-comment.md
         '@typescript-eslint/ban-ts-comment': ['error', {
-            'ts-expect-error': false,
-            'ts-ignore': false,
+            'ts-expect-error': 'allow-with-description',
+            'ts-ignore': 'allow-with-description',
             'ts-nocheck': true,
             'ts-check': true,
         }],
@@ -55,15 +56,40 @@ module.exports = {
         '@typescript-eslint/class-literal-property-style': ['error', 'fields'],
 
         // (Prise en charge TypeScript, voir parent)
+        // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/comma-dangle.md
+        'comma-dangle': ['off'],
+        '@typescript-eslint/comma-dangle': ['error', {
+            arrays: 'always-multiline',
+            enums: 'always-multiline',
+            exports: 'always-multiline',
+            functions: 'always-multiline',
+            generics: 'always-multiline',
+            imports: 'always-multiline',
+            objects: 'always-multiline',
+            tuples: 'always-multiline',
+        }],
+
+        // (Prise en charge TypeScript, voir parent)
         // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/comma-spacing.md
         'comma-spacing': ['off'],
         '@typescript-eslint/comma-spacing': ['error', { before: false, after: true }],
+
+        // - Les définitions d'objets avec membres non connus doivent être effectuées via `Record<..., ...>` (et non `{ [key: ...]: ... }`).
+        // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/consistent-indexed-object-style.md
+        '@typescript-eslint/consistent-indexed-object-style': ['error', 'record'],
 
         // - Vérifie que les assertions de type (= le casting de type) est toujours effectué via `as` ('foo' as string).
         // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/consistent-type-assertions.md
         '@typescript-eslint/consistent-type-assertions': ['error', {
             assertionStyle: 'as',
             objectLiteralTypeAssertions: 'allow',
+        }],
+
+        // - Les imports de type doivent toujouus être effectués via `import type ...`.
+        // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/consistent-type-imports.md
+        '@typescript-eslint/consistent-type-imports': ['error', {
+            prefer: 'type-imports',
+            disallowTypeAnnotations: true,
         }],
 
         // (Prise en charge TypeScript, voir parent)
@@ -77,6 +103,7 @@ module.exports = {
         '@typescript-eslint/dot-notation': ['error', {
             allowKeywords: true,
             allowPrivateClassPropertyAccess: false,
+            allowProtectedClassPropertyAccess: false,
         }],
 
         // - Requiert que les fonctions et méthodes spécifient leur type de retour.
@@ -211,10 +238,26 @@ module.exports = {
         // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-confusing-non-null-assertion.md
         '@typescript-eslint/no-confusing-non-null-assertion': ['error'],
 
+        // - Empêche l'utilisation d'expressions de type `void` dans des contextes déroutants.
+        // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-confusing-void-expression.md
+        '@typescript-eslint/no-confusing-void-expression': ['error', {
+            ignoreArrowShorthand: false,
+            ignoreVoidOperator: false,
+        }],
+
+        // - Empêche l'utilisation de `as` là ou `!` serait plus approprié.
+        // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/non-nullable-type-assertion-style.md
+        '@typescript-eslint/non-nullable-type-assertion-style': ['error'],
+
         // (Prise en charge TypeScript, voir parent)
         // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-dupe-class-members.md
         'no-dupe-class-members': ['off'],
         '@typescript-eslint/no-dupe-class-members': ['error'],
+
+        // (Prise en charge TypeScript, voir parent)
+        // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-duplicate-imports.md
+        'no-duplicate-imports': ['off'],
+        '@typescript-eslint/no-duplicate-imports': ['error'],
 
         // (Prise en charge TypeScript, voir parent)
         // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-empty-function.md
@@ -249,11 +292,19 @@ module.exports = {
 
         // - Interdit l'utilisation de `void` en dehors des retours de fonctions et des types génériques.
         // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-invalid-void-type.md
-        '@typescript-eslint/no-invalid-void-type': ['error', { allowInGenericTypeArguments: true }],
+        '@typescript-eslint/no-invalid-void-type': ['error', {
+            allowAsThisParameter: true,
+            allowInGenericTypeArguments: true,
+        }],
 
         // - Vérifie la déclaration des constructeurs dans les classes et `new()` dans les interfaces (et pas un mélange des deux).
         // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-misused-new.md
         '@typescript-eslint/no-misused-new': ['error'],
+
+        // (Prise en charge TypeScript, voir parent)
+        // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-loop-func.md
+        'no-loop-func': ['off'],
+        '@typescript-eslint/no-loop-func': ['error'],
 
         // - Interdit l'utilisation des namespaces, sauf dans les fichiers de définition.
         // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-namespace.md
@@ -285,6 +336,10 @@ module.exports = {
             typesToIgnore: [],
         }],
 
+        // - Interdit les contraintes inutiles dans les génériques.
+        // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-unnecessary-type-constraint.md
+        '@typescript-eslint/no-unnecessary-type-constraint': ['error'],
+
         // (Prise en charge TypeScript, voir parent)
         // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-loss-of-precision.md
         'no-loss-of-precision': ['off'],
@@ -308,6 +363,21 @@ module.exports = {
             args: 'after-used',
             ignoreRestSiblings: true,
             vars: 'all',
+        }],
+
+        // (Prise en charge TypeScript, voir parent)
+        // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-redeclare.md
+        'no-redeclare': 'off',
+        '@typescript-eslint/no-redeclare': ['error', {
+            ignoreDeclarationMerge: true,
+        }],
+
+        // (Prise en charge TypeScript, voir parent)
+        // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-shadow.md
+        'no-shadow': 'off',
+        '@typescript-eslint/no-shadow': ['error', {
+            ignoreTypeValueShadow: false,
+            ignoreFunctionTypeParameterNameValueShadow: false,
         }],
 
         // (Prise en charge TypeScript, voir parent)
@@ -391,6 +461,11 @@ module.exports = {
             named: 'never',
         }],
 
+        // (Prise en charge TypeScript, voir parent)
+        // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/space-infix-ops.md
+        'space-infix-ops': ['off'],
+        '@typescript-eslint/space-infix-ops': ['error', { int32Hint: false }],
+
         // - Requiert que les switchs soient exhaustifs lorsqu'ils utilisent un type `union` permettant de détecter ce genre d'erreur.
         // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/switch-exhaustiveness-check.md
         '@typescript-eslint/switch-exhaustiveness-check': ['error'],
@@ -457,6 +532,9 @@ module.exports = {
 
         // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-for-in-array.md
         '@typescript-eslint/no-for-in-array': ['off'],
+
+        // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-implicit-any-catch.md
+        '@typescript-eslint/no-implicit-any-catch': ['off'],
 
         // @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-implied-eval.md
         '@typescript-eslint/no-implied-eval': ['off'],